requirements.txt
streamlit>=1.28
tensorflow>=2.9
numpy>=1.22
Pillow>=9.3.0
scikit-learn>=1.1
matplotlib>=3.5
pandas>=1.5
opencv-python>=4.7
seaborn>=0.11